Clean up code in SentimentAnalyzer and TopicClassifier
Renamed SentimentAnalyzer to SentimentClassifier, so that both the topic and sentiment classifiers have similar names.
Made it so topic classifier now reads a saved classifier. Left in the method to create that classifier, but it's uncalled.
Sentiment and topic classifiers now throw the exceptions in their methods so that main can handle them.
Renamed classifier.txt to SentimentClassifier.txt to not cause confusion with the new TopicClassifier.txt.